Committing some of the fixes for file/project bugs.
Audacity will now move a file out of the way if it is about
to be overwritten by an export.  It should no longer be possible
to corrupt everything just be exporting.

Save As... for projects now does the right thing (copies all
associated files, so the old project is still intact).