Remove recursion while checking result type for labels as it could overflow the stack for large numbers of labels.
Reworked the way result type is checked to reduce time/cpu.  An example...to return 50 labels, over 68850 requests to xlisp were made simply to get a pointer to the label.  For 100 labels, this blossomed to over 525000(!) calls.  And this doesn't count the number of calls to xlisp that were done to validate the values of each label.