Sped up HistoryWindow updating by eliminating duplicate clearing and refilling the wxListCtrl. In the worst case, where the window had not yet been created, it would clear and refill the same content in the list 3 times. Otherwise it would always do it 2 times.
The only call to Show for HistoryWindow is in AudacityProject::OnHistory, and it always calls HistoryWindow::UpdateDisplay. All HistoryWindow::UpdateDisplay does is check IsShown() and if so, call HistoryWindow::DoUpdate. 

Removed the call to HistoryWindow::DoUpdate in the constructor because AudacityProject::OnHistory will get that done. 

Removed HistoryWindow::OnShow. No reason to implement it outside the normal wxDialog::Show() mechanism, and all it did was call HistoryWindow::UpdateDisplay (and then AudacityProject::OnHistory would do that again).

Didn't see any other reasons it has been slow, as it's just getting the state names and sizes from the stack, not actually examining anything about the states. It probably doesn't really need to clear the list and repopulate it every time, but let's see if this gives sufficient improvement before adding a lot of mechanisms for tracking exactly what needs to change.