Testing a link on a non-development machine, I wanted to make sure I had the right version of Audacity, so checked the AboutDialog. Found some problems and fixed them:
Martyn was out of alphabetical order. The reason was that the role values for Leland and Markus were not changed when they were moved to the emeritus list. 

Went ahead and split up the developer and support lists, and removed the comment, as that's how they'll appear in the interface and it's easier to alphabetize and maintain same-role lists rather than interleaved.